+++ example/electron12-vite/vite.config.ts
@@ -7,9 +7,6 @@ export default defineConfig({
       {
         entry: "electron/main.ts",
       },
-      {
-        entry: "electron/preload.ts",
-      },
     ]),
   ],
 });